separate states
